requirements.txt
asgiref==3.4.1
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
psycopg2==2.8.6
django-heroku==0.3.1
gunicorn==20.1.0
httplib2==0.19.0